requirements/default.txt
numpy==1.19.1
pandas==1.1.0
matplotlib==3.2.2
seaborn==0.11.2
statsmodels==0.11.1
ipywidgets==7.5.1
networkx==2.4
progress==1.5
sqlitedict==1.7.0
jsonpickle==2.0.0
nxviz==0.6.2
tqdm==4.60.0
sigfig==1.1.8
pympler==0.9
tabulate==0.8.7
ray==1.0.0
tree==0.2.4
dm-tree==0.1.5
opencv-python==4.4.0.2
lz4==3.1.0
gym==0.17.3
wandb==0.13.1
hydra-core==1.2.0
libtmux==0.15.8
torch
torchaudio
torchvision
